drawing on images works, and the drawing stays with the images as it should. i'm close to being ready to integrate this into the main project

+++ PaintingDemo/PaintingSample/PaintView.h
@@ -13,8 +13,7 @@
 @interface PaintView : UIView<PaintTouchViewDelegate> {
     void *cacheBitmap;
     CGContextRef cacheContext;
-    float hue;
+    CGFloat hue;
 }
-
 
 @end
